Fix CSIMC cards refusing to boot by adding a loop delay.
As far as I can tell this main loop was previously CPU-limited,
and the much faster CPU in the control machine was flooding the
token ring network with requests - denying the CSIMC cards the
ability to send the packets down the chain.

Adding a delay immediately after sending our data to the ring
gives the packets enough time to propagate to the correct card.